Add lock to protect IEffect during shutdown.
When the media server is killed AudioEffect is notified through the
binder by a call to AudioEffect::binderDied() where the client is
notified and the IEffect is cleared. It has been seen that the
destructor has been called at almost the same time from different
threads.
Timing can lead to a call to IEffect.clear() in
AudioEffect::binderDied() after the null check of IEffect in the
destructor but before the calls through IEffect leading to a null
pointer error.

Change-Id: I141d61907fa5928201abad239c3c97ceae40561e