Dont discard Play Response with valid npt1 value
If the range attribute in a Play response does not
contain a valid range, i.e. valid npt1 and npt2 values
where npt2 is higher than npt1, the Play response was
discarded. This meant that the normal playtime mapping
was not updated. After a seek operation the Play
response might for example contain a "Range: npt=28-".
This meant that the progress bar value would not be
updated after the seek operation.
This fix makes sure that the normal playtime mapping
is updated if the Play response contains a valid npt1
value in the Range attribute.

Change-Id: I8850e9957f0e8c803607b859edc34cf1fb2d5213