Playtime error after draging the progress bar to end.
The seekTime after draging the progress bar to end was found
to be more than total duration of file. As a result zero
offset was returned and playback is started from the start.
check is added for this condition and if condition met, it
will return ERROR_END_OF_STREAM.

Change-Id: Ic9e235e5724b6bd6fedfca9c0daf53dd22c0b853
Signed-off-by: Manjunath <manjunathx.b.s@intel.com>
Signed-off-by: MalimeeX <meenakshix.mali@intel.com>