Fix tombstone in AudioTrack::processAudioBuffer
This tombstone is observed in stress test, when we continuously touch the screen to repeat
the key tone. This issue can occur when the AudioTrack instance object has been deleted but the
AudioTrackThread object has not been deleted and then, when it enters its threadLoop method,
we get the segmentation fault while executing the processAudioBuffer method on the deleted
object. Therefore, we should createTrack_l first and if we succeed, we create
the AudioTrackThread to run. Otherwise tombstone happens with NULL mAudioTrack and NULL mCblk.

Change-Id: Ie3f46176ff3bb39a299dca1259406c8cdf898ca6
Signed-off-by: xueminsu <xuemin.su@intel.com>
Signed-off-by: malimeeX <meenakshix.mali@intel.com>