Three little Shelf drum on the left and gong on the above cannot sound
While finding the duration of ogg files, the last granule
position is searched. While doing this, finding EOF is
considered as an error, whereas it should be considered
as a warning and the granule position should be searched
again by moving the current pointer still back. This fixes
the issue.

Change-Id: I0fb6b23f85bf54ba512c7e82a1875c1b66cfd171
Signed-off-by: Arulselvan M <arulselvan.m@intel.com>
Signed-off-by: Jin Wei <wei.a.jin@intel.com>