Fix to play AAC file with ID3 tag.
When the file is scanned, the ID3 tag is found to have a footer at the end of audio
data. And getAdtsFrameLength returns zero after reading the footer. As a result, the
duration and mInitCheck are not set.
This fix considers the data till footer and sets the duration based on number of frames
read till the footer/corrupt frame.

Change-Id: I0eae71c0c2906a9d918b5e6edaf5a96f7a8029c7
Signed-off-by: Manjunath <manjunathx.b.s@intel.com>
Signed-off-by: gbhakteX <gurudattax.bhakte@intel.com>