Fix to solve seek to the last buffer issue for an offloaded song.
When seeked in the last buffer for an offloaded song, the data from seek position
to EOS is played after moving to the next song. This is because sink is stopped
and drain is issued before the last seeked data from flinger is rendered. There
is no write call to the HAL after seek. The OffloadThread is always in sleep since
prepareTracks does not return MIXER_STATUS_READY.
With this fix, if EOS is reached, querying the HW for render position is removed
and the old PositionRealTime is used.

Change-Id: Ibb1e40ca1e06f700429aa7bd2bdc6bbf2937a119
Signed-off-by: jpadmana <jayashree.r.padmanaban@intel.com>
Signed-off-by: gbhakteX <gurudattax.bhakte@intel.com>