Fix to solve the issue of different songs getting played on HDMI and DUT during Playback of mp3 files on HDMI.
For shorter mp3 files, during offload and End of stream in AudioTrack, if a HDMI device
is connected, teardown doesn't happen. The MEDIA_PLAYBACK_COMPLETE gets posted before
the teardown event. This leads to jump of current song to next song in the playlist.
This fix makes sure that teardown is posted so that the current song is continued on HDMI.

Change-Id: I8d422c74d1b8d343dcc3236e8eb44626760d1d33
Signed-off-by: samaga Prasanna Krishna <samaga.krishna@intel.com>
Signed-off-by: gbhakteX <gurudattax.bhakte@intel.com>