Fix to correct "timeout" variable declaration, to improve audio quality.
"struct timespec timeout" in obtainBuffer is declared as a local variable in AudioRecord
and AudioTrack. This impacts the ClientProxy::obtainBuffer() function in
AudioTrackShared.cpp where "measure" variable will not be set to true leading us to skip
the shared memory futex. This leads to a poor audio quality with cuts depending on the
values of "tv_sec" and "tv_nsec" which are unknown before execution.
Moving the declaration of "timeout" outside of the 'if' condition solves the issue.

Change-Id: I73ceb081cf129fb50f93095d3b97d9b80b7151bb
Signed-off-by: Julien Delayen <julien.delayen@intel.com>
Signed-off-by: gbhakteX <gurudattax.bhakte@intel.com>