camera:Fix for assigning proper IDs to O/P Streams
The output streams were assigned ID's in incremental fashion for
every creation. There was no account kept for the ID's. Since the data
type for the ID's was char, it caused the Camera App to crash after 64
loops of Camera Mode Switch. Fixed the issue by keeping a count of the
used ID's and re-using the ID's once a stream is deleted.

Change-Id: Ib1956d4124b4b4e5470d96b97e3123f2121b1e60