Offload: Progress time is incorrect while seeking MP3 in .mka container without cues
In case of music offload, when a MP3 content in .mka container format
without cues/seek information is played, seek does not work properly.
During offload, when seek is performed on such files, even after reaching
end of file, the audio continues to play erroneously.
To fix this, a check is added to allow seek operation on files only if seek
information is present. With this fix offload seek behaviour is same as
non offload seek.

Change-Id: If77e541c2e2d26768bb449613a3e838191c58811
Signed-off-by: Divya Prakash <divya1.prakash@intel.com>
Signed-off-by: Mazhar <mazharx.sameullah@intel.com>