Fix crash in H263 encoder on 64 bit platforms
The HTFMPrepareCurMB function initialises a 256 byte buffer in the
VideoEncData structure.  This buffer, VideoEncData.currYMB, is
initialised by a for loop that attempts to write 4 32 bit words 16 times.
The problem is that it writes these words using a Long* and so on 64 bit
platforms the code ends up writing 512 bytes and not 256.  The result
is that other VideoEncData members get overwritten and the encoder crashes
later on in EncodeFrameCombinedMode.

Change-Id: I5387ad5bc9a85c587a6b93c7ed5795ebf3d6917b
Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>