[WIP] Add libcrypto for users of libstagefright.
libstagefright uses some MD5 functions that used to be in libc.
Originally, this code was also in bionic, but it was removed
recently. As an initial fix, I statically linked in the MD5 functions
into the library. I think it's better to modify all of the users of
the library to link against the shared libcrypto library just in case
we find a bug in the MD5 implementation.

Change-Id: I62dffe46717a1188a0afbf4500e658b389220ba5