Camera HAL3/API1: fix to have zsl receive unique frames
ZslProcessor's update stream is called multiple times, which causes it to
register it as a listener with FrameProcessorBase multiple times. This causes
the FrameProcessorBase to send one frame to the ZslProcessor multiple times
resulting in flooding of the Zsl buffer queue with same frames.

This fix prevents re-registration of an existing client.

Change-Id: I9c88e91d716c829cc868ab3663108d223407dd60