MediaCodec: Properly handle cases where ACodec spontaneously shuts down
If a codec fails to allocate the output port buffers, it will
transition to the loaded state, without the caller knowing
about this.

When this happens, MediaCodec will get a kWhatShutdownCompleted
message, but the MediaCoec object is not in the expected
state. Therefore, MediaCodec needs to know that this is a
shutdown that it itself didn't initiate, waiving the
check for the MediaCodec state.

If the caller calls MediaCodec.stop() quickly after the error
is reported, MediaCodec might transition to the RELEASING
state before ACodec has sent the spontaneously triggered
kWhatShutdownCompleted message. In this case, MediaCodec will
still request the codec to shut down, resulting in two
kWhatShutdownCompleted messages. When the latter is received
we are already in the UNINITIALIZED state - keep a flag
indicating that the codec has shut down spontaneously and that
inconsistent state in the kWhatShutdownCompleted handler is
expected.

Likewise within ACodec; if the caller tries to call stop()
on the MediaCodec object before ACodec has indicated that it
has shut down, it will send a kWhatShutdown message with
keepComponentAllocated=1 even though ACodec already is in
UninitializedState. Ignore the mismatched keepComponentAllocated
flag in this case, if states are out of sync due to the
spontaneous shutdown.

Change-Id: Ic1fd52e98259a2530095389e60d4f06cb19a09d1