libstagefright: fix u/v and rgb order issue in nv12 color convertor
Issue is that color in video thumbnamil is different with it in real
video play. It should come from color convert for YUV420 semi planar.

According to omx IL spec, for format "OMX_COLOR_FormatYUV420SemiPlanar",
u and v pixels are interleaved with the first u value first, that is,
NV12 format. In original convertor, v is sampled firstly.
Another issue is it has wrong RGB array order - RGB data is saved as
"BGR".
With u/v and rgb order change, NV12 thumbnail color is same with it
in video playing.

Change-Id: I8317db6f5f89f1ab34c335a519a542156fd6d843

Signed-off-by: Junxiao Chang <junxiao.chang@intel.com>
Signed-off-by: Hongfu Ruan <hongfu.ruan@intel.com>