Update the video track's RTP Anchor value with the first AU's timestamp
Corrupted frame is observed at the first second during test below:
CTS Verifier 4.4_r3
Category
- Streaming Video Quality Verifier
  - RTSP
Test
- H263 Video, AMR Audio
- MPEG4 SP Video, AAC Audio
- H264 Base Video, AAC Audio

The RTSP server sends the first SR with the RTP timestamp value bigger than
the real first access unit's RTP timestamp and thus it is taken as an early
coming packet and dropped. Since the first AU is a key frame, it will cause
the corruption you see.

This patch is created to prove the analysis above. The corruption is gone
after this patch is applied.

It should be a defect of Android, but it's kind of corner case since normally
the server should give a correct RTP timestamp in SR.

Change-Id: I34bb02f7eceea61ca10dc2811ea2b0e96eae0bbb
Signed-off-by: Dan Liang <dan.liang@intel.com>
Signed-off-by: Guobin Zhang <guobin.zhang@intel.com>