Enable WiFi Display UIBC feature for Source device
Follow the WiFi Display Specification, UIBC (user input
back channel) is negotiated in M3 stage of RTSP and it
use TCP to communicate with Sink.
The UIBC data will send to WifiDisplayController for
injecting the event to framework.
As the resolution might be different between source device
and sink device so before transmiting the coordinate it
should be calculated according to the parameters.
When source device rotated the coordinate will change but it
won't happen in sink device. So check the rotation when UIBC
packets arrived.

Change-Id: Ifb30f7991f45517f0061279c8ba05674cb07e353
Signed-off-by: Wang Haoran <Haoran.Wang@freescale.com>
Signed-off-by: guoyin.chen <guoyin.chen@freescale.com>