Enable WiFi Display UIBC feature for Source device
Follow the WiFi Display Specification, UIBC (user input
back channel) is negotiated in M3 stage of RTSP and it
use TCP to communicate with Sink.
The UIBC data will send to WifiDisplayController for
injecting the event to framework.
As the resolution might be different between source device
and sink device so before transmiting the coordinate it
should be calculated according to the parameters.
When source device rotated the coordinate will change but it
won't happen in sink device. So check the rotation when UIBC
packets arrived.

Change-Id: I612e296f1855a3ddb4f48d955ea728d43354fcd6
Signed-off-by: Wang Haoran <Haoran.Wang@freescale.com>
Signed-off-by: guoyin.chen <guoyin.chen@freescale.com>