Add missing #includes
Some header files in stagefright try to use
DISALLOW_EVIL_CONSTRUCTORS (and even DISALLOW_IMPLICIT_CONSTRUCTORS,
which isn't defined anywhere in stagefright) without including the
proper header.

With clang, this results in a build failure - with gcc, it results in
DISALLOW_EVIL_CONSTRUCTORS/DISALLOW_IMPLICIT_CONSTRUCTORS not doing what
they were intended to do, instead appearing as a method in the DWARF
debug sections of files using the headers.

Change-Id: Ie4e99affe98ed8fb1bfcf5ec604c19dcd5311c62
Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>