Fix struct vs. class mismatches
Fix class vs. struct mismatches clang whines about.

Since there doesn't seem to be a particular pattern to when the code
uses struct vs. class (maybe it started out as C code and everything
that existed in C times remained a struct), I'm following the idea that
anything that has methods is a class.

Another option might be to just use struct everywhere, that would
probably be the smallest number of changes.

Change-Id: Ia551c796b6a5df8317943715447d0c28cdbf465a
Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>