fix: memory leak during audio routing in audiopolicy
createAudioPatch is called for changing device in setOutputDevice.
After calling it, AudioPatch instance is allocated in the heap
using patchHandle returned.

However, if createAudioPatch was called with delayMs then patchHandle
returned is 0. Because AudioCommend in Audiopolicyservice is designed
to return instantly if delayMs isn't 0.
Overall, new AudioPatch instance will be created over and over. This
causes memory leak in mediaserver.

Finally, this is temporally fix. because routing with delay should
work. But, this is very critical issue I think, and there are no
time no idea to fix this quickly. We need to fix this for
the next one.

Change-Id: I067fafca9c3234293e2699783609c66e9d6bae67