Use correct start time values at HLS bandwidth switch
When performing a bandwidth change from audio only to audio/video
then is the latest enqueued meta for video at least one segment
behind audio (if the video recovers quickly) and this results
in a faulty startTimeUs value being set in LiveSession.

This commit takes the values from the source with the highest
segment start time or if the segment start times are equal then
the lowest startTimeUs is chosen.

Change-Id: I14cf1186d0daf517a24e8423c3a708b4c9ba06c4