Improve bandwidth selection algorithm for HLS
This fix includes a new algorithm that estimates the
current bandwidth using weighted values of segment
download times. The values are weighted depending on
how long time ago the segment was downloaded.
Also a maximum of 61 seconds of playtime data is used
for the estimation.

Do not switch to a lower BW index if current buffer
duration exceeds two target durations.

Change-Id: I49decf26d358211fe5f9863335f0710609c934f5