Camera2Client: optionally defer initialization of JPEG stream
In a HAL3 device, it is possible to have per-resolution frame rates
declared.  A camera might expose a 15 fps full resolution capture
mode, yet still declare 30 fps preview modes.  The legacy API doesn't
understand this idea, so Camera2Client always configures the specified
"slow" capture stream underneath startPreview().

This leads to failures in CTS's CameraGLTest, which attempts to
startPreview() on enumerated modes, then gets confused when the FPS is
not as declared due to the unused (and unrepresentable!) "hidden
capture mode".

As this initialization of the JPEG stream is a pure optimization,
allow devices to disable it with a property so that direct measurement
of preview FPS in isolation will work correctly.  The API mismatch
will only become apparent when frame rate drops following the first
takePicture().

Change-Id: I28433d3a186fec43662f938dc575f5c3000d1e76
Signed-off-by: Andy Ross <andrew.j.ross@intel.com>