Improve bandwidth changes for HLS
When switching bitrate, the new fetcher would download a
segment in the future on the new bitrate. Then the old
fetcher would be resumed and continue downloading on
the old bitrate until the time where the new segment
started. This would likely cause buffering at bandwidth
decrease.

This patch instead calculates which segment should be
downloaded on the new bitrate and then resumes or stops
the old fetcher before the download is started on the new
fetcher. This way buffering can be avoided.

Change-Id: I918c80947339d951e8c1ed52cadb474f0d466146