Implement ReconfigurationState in LiveSession
Reconfiguration can be triggered by select track, seek and
bandwidth changes. Currently it'ss only a boolean variable used
to handle these reconfigurations but that is not always enough.

This commit introduces a state variable with the values:
        -NONE
        -RECONFIGURATION_PENDING
        -RECONFIGURATION_IN_PROGRESS
        -DISCONNECT

It will now be much easier to allow or reject the reconfiguration
depending on which state previous rencofiguration is in.

Change-Id: I52f31fde1cf3f09f3cb4152c45e4e86e1bb3f7c4