MediaServer crash due to unsupported container format
MediaServer crash due to unsupported container format via RTSP
in ASessionDescription::getFormatType()

The crash is because of parsing error due to improper rtpmap
attribute set by RTSP server. To avoid this, check if the number
of rtpmap attributes and media-formats obtained as response to
DESCRIBE request are equal. If unequal results in parsing error.

Change-Id: I70d5466cd0c51b967ca301c378842bb0e50b7c4a
Signed-off-by: samaga Prasanna Krishna <samaga.krishna@intel.com>
Signed-off-by: Yong Yao <yong.yao@intel.com>