Gapless meta data to be passed to HAL for non gapless files.
If files which do not have gapless meta data, paddingSamples
and delaySamples are not passed to HAL. This results in
incorrect progress bar update. This results progress bar
stuck in last position and the progress bar timer will
keep updating in music player during playback.

The gapless meta data values for paddingSample and delaySamples
need to be passed to HAL even if the file does have it. The
code changes are to address this.

Change-Id: Ia59965d9e4a1f28e8be3c97a6dc083f5ab2d151a
Signed-off-by: samaga Prasanna Krishna <samaga.krishna@intel.com>
Signed-off-by: Yong Yao <yong.yao@intel.com>