Remove dependency from mediaplayerservice start
Currently there is following list of dependencies:

* batterystats service is started in system_server process.

* mediaplayerservice service start time depends on batterystats service
start resulting "media.player" registration delay until batterystats
service is registered.

This is to handle the degrade introduced by commit
0b30fd4ccf5218b15f770e778703d69cd3432446

Moving the parts that depend on batterystats into a background
thread, that will run at a later time not delying initialization
of mediaplayerservice.

Change-Id: Ibc24d5b477942991e6a0a65034db7d381c92d62a