Enhance code robustness of AudioSystem
Functions getOutputSamplingRate, getOutputFrameCount and
getOutputLatency will call getOutput without releaseOutput.

In case getOutput is called with a stream type that is routed to
a device with the AUDIO_OUTPUT_FLAG_DIRECT flag enabled then a new
output object will be created in the AudioPolicyManager.

The AudioPolicyManager will ensure that any existing direct
outputs are reused or replaced and therefore this implementation
is just sub-optimal because it might leave unused objects allocated
in the system.

The solution for this issue is to change the contract for
getOutput() in AudioPolicyManager implementation and not increment
the ref count on the output stream in getOutput() but only in
getOutputForAttr().

Change-Id: I917f3f232759a6b8e116ca47f53a6a518a1ffdf5