Enhance code robustness of AudioSystem
Functions getOutputSamplingRate, getOutputFrameCount and
getOutputLatency will call getOutput without releaseOutput.

In case getOutput is called with a stream type that is routed to
a device with the AUDIO_OUTPUT_FLAG_DIRECT flag enabled then a new
output object will be created in the AudioPolicyManager.

Depending on the product/vendor AudioPolicyManager implementation
two things can happen:

On the best case scenario (also default AOSP implementation),
the AudioPolicyManager will ensure that any existing direct
outputs are reused or replaced and therefore this implementation
is just sub-optimal because it leaves unused objects allocated in
the system.

On the worst case scenario, the AudioPolicyManager will allocate a
new direct output every time getOutput is called and will not clean
up any existing direct output from previous calls.

Change-Id: I917f3f232759a6b8e116ca47f53a6a518a1ffdf5