Fix timestamp anchor problems for RTSP
Correctly set the media anchors for RTSP when seeking, pausing
and starting. Both when using sender reports or when using
fake timestamps. The timestamps are used when synchronizing the
audio and video streams.

The timestamps calculated in MyHandler using RTSP Sender Reports were
overwritten at a later stage in RTSPSource. This commit also removes
the code that writes over the existing timestamps and instead uses
the timestamps calculated in MyHandler.

Change-Id: Idd238df5ebeef2d2a7da03730cb4a70445a2339d