Bug 179658: AMediaCodec_createCodecByName() now returns NULL if an error occurred.
Calls to the 'create functions' (AMediaCodec_CreateCodecByName(), AMediaCodec_createDecoderByType(), and AMediaCodec_createEncoderByType()) currently return non-NULL values when an error occurs. The callee has no way to determine if an error occurred. If AMediaCodec_configure() is called with the non-NULL AMediaCodec* from a failed call to one of the create functions, it will crash. In NdkMediaCodec.cpp, the createAMediaCodec() method accepts an output status_t* parameter, but the create functions don't use it. This patch uses the status_t parameter and returns NULL to the caller when an error occurs. Ideally we would pass the error code back to the user, but this communicates the error state without modifying the public API.

Change-Id: Ifafc881b2f61702bf903bbadf991dd26a98b1489