Audio Effects are not processing first buffer
Current implementation in AudioFlinger set the effects to state IDLE
when they are created. Later on when the effect is enabled by the client,
the state changes to STARTING.

Then when the audio playback starts, the AudioFlinger thread loop calls
EffectChain::process_l() to perform the effects processing. However this
method will first call process() and then updateState(), so the first
process won't do anything because the effect is in STARTING state. After
the call to updateState, the state is moved to ACTIVE and then next calls
to process_l() will work as expected.

Change-Id: I9dfd3d5a0e53403034eb42f9366e1b3cdc5249c0