Improve error handling in ABitReader
The ABitReader class relies on an ASSERT-style CHECK_GT macro to detect errors.
Since these errors can be caused by untrusted input, crashing the process
hosting libstagefright should be avoided if at all possible. This patch adds
return values for the getBits and fillReservoir methods. Further, it modifies
all callers to pay attention to the indication there was an error and bail
gracefully.

Change-Id: I60c73571d26608c416fa4aea46069f0687de52f7