Decoder already exists message shown as info message
"Decoder already exists message" is an error message shown pretty frequently when playing video/audio content. This happens everytime the system needs to instantiate a decoder that has both a hardware based and a software based implementation. For example, this happens with AAC decoder.

This is not really an error so I think this message should be logged just as information.

Change-Id: I0d5d03df66f20fe7b90e7aa02743715357f082ff
Signed-off-by: joli <jeoliva@gmail.com>